Changed errand templates to enable different rendering methods
Removed the assumption that what we are rendering to is erb, so
that it becomes easier to render to .sh for the purpose of running
the errands outside of Ruby (as in test-errand).

This is a partial solution to issue #134, but test-errand rendering
is incomplete, since it is missing some of the context that is
usually inserted by build(). Specifically, the `is_app` and other
similar flags, but there are likely more.