feat: add way for skipping exporting specific properties into the (#326)
environment

With lengthy environment variables the deploy-all and delete-all jobs
reach a limit in the operating system when exporting values and all
subsequent bash commands fail with `Argument list too long` message.

The limit when setting env variables into the app is reached before this
passes, and we have successfully been able to workaround this by
providing a pre-deploy hook. However environment variables are
nevertheless exported inside the import_opsmgr_variables function and
there is no way to override this behavior

This change allows to define a `skip_export` attribute to any property
to skip exporting as environment variable both in deploy-all and
delete-all job templates. If the attribute is not specified, the
behavior remains as always.

The variables will still be set to the app with `cf set-env` as the
function that does this does not look into the environment.